application/requirements.txt
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.1
pluggy==1.0.0
pycryptodome==3.17
pydantic==1.10.7
pytest==7.3.1
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
